# docs/openapi-v1-incremental.yaml
openapi: 3.1.0
info:
  title: 淘到宝引擎 V1 增量契约
  version: 1.1.0
  description: >-
    本文件只描述相对冻结 V0.5 openapi.yaml 新增的接口；冻结路径和字段语义保持不变。
servers:
  - url: /api/v1
paths:
...
  /auth/invitation/verify:
    post:
      operationId: verifyInvitation
      description: >-
        Verifies either a server-registered short code in the tdb-xxxxx format
        or a backward-compatible signed invitation. Registered codes and new
        signed tokens may declare 1-100 total redemptions; legacy tokens default
        to one. Every successful redemption still creates a distinct,
        short-lived, single-use OAuth proof.
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [invitation_code]
              properties:
                invitation_code: {type: string, minLength: 1, maxLength: 256}
      responses:
        '200': {description: 签名邀请码验证成功、令牌摘要完成服务端登记，并为当前使用者建立限时且单次使用的 HttpOnly Cookie}
        '401': {description: 邀请码签名无效、已过期、超过最长有效期或共享兑换额度已用完；响应不会泄露具体原因或正确值}
  /auth/demo/start:
    post:
      operationId: startSharedDemoSession
      description: >-
        消费当前浏览器中已验证且单次有效的邀请码证明，为评委创建独立体验身份，
        并进入固定的共享比赛演示工作区。该接口不执行或伪装跨租户飞书 OAuth，
        也不会授予访问任何个人飞书云盘、通讯录或多维表格的权限。
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      responses:
        '200':
          description: 已建立共享 Demo 会话并写入 HttpOnly Session Cookie
        '401':
          description: 邀请码证明无效、过期或已被消费
...
components:
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 8, maxLength: 128}
...